resume-scanner-backend/requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
spacy
scikit-learn